% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants}
\alias{constants}
\alias{MOS}
\alias{MONTHS}
\alias{current_month}
\alias{current_year}
\alias{data_urls}
\alias{common_columns}
\title{Package Constants}
\format{
Various.
}
\usage{
MOS

MONTHS

current_month

current_year

data_urls

common_columns
}
\description{
Package Constants
}
