% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{get_climate_series_names}
\alias{get_climate_series_names}
\title{Get Names of Climate Time Series from Data Set}
\usage{
get_climate_series_names(x, conf_int = FALSE, invert = TRUE)
}
\arguments{
\item{x}{A data frame (down)loaded by function \code{\link{get_climate_data}}.}

\item{conf_int}{Logical; if \code{FALSE}, omit columns containing confidence intervals.}
}
\value{
A character vector of column names.
}
\description{
Extracts column names from downloaded climate data that correspond specifically to climatological time series.
}

