% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{\%_\%}
\alias{\%_\%}
\title{Concatenate Strings Easily}
\usage{
a \%_\% b
}
\arguments{
\item{a}{R object to be converted to a character vector.}

\item{b}{R object to be converted to a character vector.}

\item{sep}{A character string to separate the terms; passed to \code{\link[base]{paste}()}.}
}
\value{
The concatenation of \code{a} and \code{b}.
}
\description{
Allows quick chaining together of character strings.
}
\examples{
who <- "world"
"Hello " \%_\% who \%_\% "!"

}
\seealso{
\code{\link[base]{paste}}
}

