% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{make_time_series_from_anomalies}
\alias{make_time_series_from_anomalies}
\title{Make Time Series Objects from Climate Data}
\usage{
make_time_series_from_anomalies(x, frequency = 12L, offset = 0.5, ...)
}
\arguments{
\item{x}{A data set (down)loaded by function \code{\link{get_climate_data}}.}

\item{frequency}{The number of observations per unit of time (e.g. no. of months per year for monthly temperature data); passed to \code{link[stats]{ts}}.}

\item{...}{Passed to \code{\link{get_climate_series_names}}.}
}
\value{
An object of class \code{\link[stats]{ts}}.
}
\description{
Creates \code{ts} objects from downloaded climate data.
}
