% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{recenter_anomalies}
\alias{recenter_anomalies}
\title{Recenter Climate-Series Anomalies on a Different Baseline}
\usage{
recenter_anomalies(x, baseline = defaultBaseline, digits = 4L,
  by_month = TRUE, return_baselines_only = FALSE, ...)
}
\arguments{
\item{x}{A data frame (down)loaded by function \code{\link{get_climate_data}}.}

\item{baseline}{An integer year or, more typically, range of years on which the temperature anomalies will be centered. If \code{NULL}, no baseline centering is done; if \code{TRUE}, the default baseline of 1981–2010 is used.}

\item{digits}{An integer indicating the number of decimal places to be used; passed to \code{round()}.}

\item{by_month}{Logical; use \code{TRUE} for monthly data, \code{FALSE} for yearly data.}

\item{...}{Passed to \code{\link{get_climate_series_names}}.}
}
\value{
The data frame argument \code{x} recentered on \code{baseline}.
}
\description{
Centers or recenters climate time series on a range of baseline years; the data can already have been centered on a baseline, or can be non-centered to start with.
}

