% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constants}
\alias{MONTHS}
\alias{MOS}
\alias{constants}
\alias{current_month}
\alias{current_year}
\title{Package Constants}
\format{Various.}
\usage{
MOS

MONTHS

current_month

current_year
}
\description{
Package Constants
}
\keyword{datasets}

