% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constants}
\alias{constants}
\alias{MOS}
\alias{MONTHS}
\alias{current_month}
\alias{current_year}
\alias{data_urls}
\alias{common_columns}
\title{Package Constants}
\format{
Various.

An object of class \code{character} of length 12.

An object of class \code{character} of length 12.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{list} of length 298.

An object of class \code{character} of length 4.
}
\usage{
MOS

MONTHS

current_month

current_year

data_urls

common_columns
}
\description{
Package Constants
}
\keyword{datasets}
