% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{get_climate_data}
\alias{get_climate_data}
\title{Get Monthly Climatological Data from a Local Data Set or from the Internet}
\usage{
get_climate_data(download, data_dir = climeseries::dataDir,
  filename_base = climeseries::filenameBase,
  urls = climeseries::instrumentalUrls, baseline = 1981:2010,
  verbose = TRUE)
}
\arguments{
\item{download}{Logical: if \code{TRUE}, download the latest instrumental data from the Internet; if \code{FALSE}, load the most recent instrumental data set from the default data directory.}

\item{data_dir}{The path of the primary directory for storing downloaded instrumental temperature data. The default is a global value defined in the file "constants.R".}

\item{filename_base}{The starting filename for climate-series data files, to which "raw_" and/or a date representation will be appended.}

\item{urls}{A list of named URLs pointing to instrumental data online; the default is a global value defined in the package file "constants.R".}

\item{baseline}{An integer year or, more typically, range of years on which the temperature anomalies will be centered. If \code{NULL}, no baseline centering is done, and the string "raw_" will be appended to \code{filename_base} before the instrumental data is saved or loaded.}

\item{verbose}{Logical; passed to \code{DownloadInstrumentalData}.}
}
\value{
A data frame containing monthly climatological data sets with the following columns:
  \item{year}{The year CE.}
  \item{met_year}{The "meteorological year" starting the previous December.}
  \item{yr_part}{Monthly midpoint as a fraction of a year, i.e. (month - 0.5)/12.}
  \item{month}{Month of the year as an integer value.}
  \item{BEST}{Berkeley Earth Surface Temperature (BEST) global average combined land+SST temperature anomaly.}
  \item{BEST_uncertainty}{BEST 95\% confidence interval of uncertainty in the temperature anomaly.}
  \item{Cowtan & Way Hybrid}{Cowtan & Way hybrid reconstruction of HadCRUT4 global average combined land+SST temperature anomaly. (See \url{dx.doi.org/10.1002/qj.2297}.)}
  \item{Cowtan & Way Hybrid_uncertainty}{Cowtan & Way hybrid 95\% confidence interval of uncertainty in the temperature anomaly.}
  \item{GISTEMP}{Goddard Institute for Space Studies (GISS) global average combined land+SST temperature anomaly.}
  \item{GISTEMP NH}{GISS northern hemisphere average combined land+SST temperature anomaly.}
  \item{GISTEMP SH}{GISS southern hemisphere average combined land+SST temperature anomaly.}
  \item{HadCRUT3}{UK Met Office Hadley Centre global average combined land+SST temperature anomaly (deprecated).}
  \item{HadCRUT3v}{UK Met Office Hadley Centre variance-adjusted global average combined land+SST temperature anomaly (deprecated).}
  \item{HadCRUT4}{UK Met Office Hadley Centre global average combined land+SST temperature anomaly.}
  \item{JMA}{Japan Meteorological Agency (JMA) global average combined land+SST temperature anomaly.}
  \item{Keeling}{Scripps Institution of Oceanography CO2 measurements at Mauna Loa.}
  \item{NCEI}{U.S. National Centers for Environmental Information (NCEI) global average combined land+SST temperature anomaly.}
  \item{NCEP Surface Air}{National Centers for Environmental Prediction (NCEP) NCEP/NCAR reanalysis estimate of global average surface air temperature.}
  \item{NCEP Surface Air NH}{NCEP/NCAR reanalysis estimate of northern hemisphere average surface air temperature.}
  \item{NCEP Surface Air SH}{NCEP/NCAR reanalysis estimate of southern hemisphere average surface air temperature.}
  \item{RATPAC-A 850-300 mb}{NOAA's Radiosonde Atmospheric Temperature Products for Assessing Climate (RATPAC) global average tropospheric temperature anomaly (850–300 mb).}
  \item{RSS TLT 3.3}{Remote Sensing Systems (RSS) Temperature Lower Troposphere (TLT) global average temperature anomaly (v. 3.3).}
  \item{RSS TMT 3.3}{RSS Temperature Middle Troposphere (TMT) global average temperature anomaly (v. 3.3).}
  \item{RSS TMT 4.0}{RSS Temperature Middle Troposphere (TMT) global average temperature anomaly (v. 4.0).}
  \item{UAH TLT 5.6}{University of Alabama in Huntsville (UAH) Temperature Lower Troposphere (TLT) global average temperature anomaly (v. 5.6).}
  \item{UAH TLT 6.0}{University of Alabama in Huntsville (UAH) Temperature Lower Troposphere (TLT) global average temperature anomaly (v. 6.0).}
}
\description{
Loads climatological data from a saved data set or downloads and saves it directly from the Internet.
}
\examples{
\dontrun{
## Download both centered and "raw" data.

d <- get_climate_data(download=TRUE, baseline=TRUE)
e <- get_climate_data(download=TRUE, baseline=FALSE)

## Load both centered and "raw" data.

d <- get_climate_data(download=FALSE, baseline=TRUE)
e <- get_climate_data(download=FALSE, baseline=FALSE)

## Which year is the warmest?

inst <- get_climate_data(download=FALSE, baseline=TRUE)
series <- setdiff(names(inst), c(commonColumns, c("Keeling")))
yearType <- "year" # "year" or "met_year" = meteorological year.
annual <- sapply(series, function(s) { rv = tapply(inst[[s]], inst[[yearType]], mean, na.rm=TRUE); rv = rv[!is.nan(rv)]; rv })
lastYear <- as.integer(format(Sys.Date(), "\%Y")) - 1

## How many months for 'lastYear' have data?
sapply(inst[inst[[yearType]] \%in\% lastYear, series], function(s) sum(!is.na(s)))
mapply(function(x, y) x[y][1L], annual, sapply(annual, function(s) { order(s, decreasing=TRUE) }))

## Calculate max. anomaly for years not in 'excludeDate'. (Allows exclusion of e.g. partial current year.)
excludeDate <- as.integer(format(Sys.Date(), "\%Y")) # Or excludeDate <- c(2016)
annualLt <- sapply(annual, function(s) { s[!(as.numeric(names(s)) \%in\% excludeDate)] })
mapply(function(x, y) x[y][1L], annualLt, sapply(annualLt, function(s) { order(s, decreasing=TRUE) }))
}

}

