% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-series.R
\name{plot_models_and_climate_data}
\alias{plot_models_and_climate_data}
\title{Plot Climatological Time Series with Model Results}
\usage{
plot_models_and_climate_data(instrumental, models, series = NULL,
  scenario = NULL, start = 1880, end = NULL, ma = NULL, ma_i = ma,
  baseline = NULL, ylim = c(-1, 1), center_fun = "mean",
  smooth_center = FALSE, envelope_coverage = 0.95,
  envelope_type = c("quantiles", "range", "normal"), plot_envelope = TRUE,
  smooth_envelope = TRUE, col_m = NULL, col_m_mean = NULL,
  alpha_envelope = 0.2, envelope_text = "model coverage",
  legend... = list(), plot_i... = list(),
  col_i_fun = RColorBrewer::brewer.pal, col_i_fun... = list(name =
  "Paired"), alpha_i = 0.5, conf_int_i = FALSE, ci_alpha_i = 0.3,
  omit_series = c("Keeling"), ...)
}
\arguments{
\item{instrumental}{An instrumental-temperature dataset (down)loaded by function \code{\link{get_climate_data}}.}

\item{models}{A modeled-temperature dataset loaded by function \code{\link{get_models_data}}.}

\item{series}{A vector containing the column names of the instrumental temperature series to be plotted. If \code{NULL}, all the series in \code{instrumental} are plotted; if \code{NA}, plotting of the instrumental temperature series is suppressed.}

\item{scenario}{A character vector corresponding to specific radiative-forcing pathways represented in subsets of the modeled temperature data. If \code{NULL}, all pathways are included. The options for CMIP3+ data are any subset of \code{c("20C3M", "SRES B1", "SRES A1B", "SRES A2")}; options for CMIP5 include any subset of \code{c("RCP 2.6", "RCP 4.5", "RCP 6.0", "RCP 8.5")}.}

\item{start}{and \code{end} are integer values of the starting and ending years of the plot, respectively; if either is given as \code{NULL}, the minimum or maximum date in \code{models} is used instead.}

\item{ma}{and \code{ma_i} are the sizes in months of the moving averages for smoothing the modeled and instrumental temperature series, respectively. If \code{ma_i} is not given, its value is the same as \code{ma}. A value of \code{NULL} means that no moving-average smoothing is done.}

\item{baseline}{An integer year or, more typically, range of years on which the temperature anomalies will be centered. If \code{NULL}, no baseline centering is done, and the "baseline" attribute of \code{models} (if it exists) is used instead.}

\item{ylim}{The y limits of the plot, i.e. the plot range of the temperature anomalies.}

\item{center_fun}{The function used to calculate the central tendency of the model runs at each time point; the default of \code{"mean"} is usually sufficient.}

\item{smooth_center}{Logical; if \code{TRUE}, LOESS-smooth the plot of the models' central tendency.}

\item{envelope_coverage}{Sets the "coverage" of the models-envelope calculated as \code{envelope_type}; e.g. for \code{envelope_coverage = 0.95} and \code{envelope_type = "range"}, the envelope covers 95\% of the difference between the maximum model value and the minimum, for each time point.}

\item{envelope_type}{A character vector matching one of several methods (e.g. "quantiles", "range", "normal") for calculating the \code{envelope_coverage} envelope characterizing the spread of the models. The default value is \code{"quantiles"}.}

\item{plot_envelope}{Logical; if \code{TRUE}, plot the envelope of 95\% model coverage.}

\item{smooth_envelope}{Logical; if \code{TRUE}, LOESS-smooth the plot of the envelope of 95\% model coverage.}

\item{col_m}{A vector of colors for each model scenario. If \code{NULL} (the default), a series of grays is used; if a single value, \code{\link{vary_brightness}} is called with that value to get a range of colors based on it.}

\item{col_m_mean}{The color of the central-tendency line for all the included model runs. The default of \code{NULL} causes a darkish gray to be used; ; if a single value, \code{\link{vary_brightness}} is called with that value to get a range of colors based on it.}

\item{alpha_envelope}{Sets the color transparency of the model envelopes to a new level in [0, 1]. If \code{alpha_envelope} is \code{NA}, existing alpha values are preserved. If \code{alpha_envelope} is \code{NULL}, the model envelopes are not color-filled.}

\item{legend...}{Takes a list of arguments to be passed to function \code{\link[graphics]{legend}}, possibly overriding default values.}

\item{plot_i...}{Takes a list of arguments to be passed to function \code{\link[zoo]{plot.zoo}}, possibly overriding default values, for the instrumental series.}

\item{col_i_fun}{A color palette for \code{series} is generated from the function value of \code{col_i_fun}. Note that the colors created for the instrumental series here can be entirely overridden by including a \code{col} argument in \code{plot_i...}; or, better, by specifying a function like \code{function(...) "red"} (which also affects the legend).}

\item{col_i_fun...}{Takes a list of arguments to be passed to function \code{col_i_fun}, possibly overriding default values. E.g. to get colors from function \code{\link[RColorBrewer]{brewer.pal}}, use \code{col_i_fun = RColorBrewer::brewer.pal} and \code{col_i_fun... = list(name="Paired")}.}

\item{alpha_i}{Sets the color transparency of the temperature series to a new level in [0, 1]. If \code{alpha} is \code{NA}, existing alpha values are preserved.}

\item{conf_int_i}{Logical; if \code{TRUE}, plot 95\% confidence intervals for those series with CI data.}

\item{ci_alpha_i}{Sets the color transparency of the temperature-series CIs to a new level in [0, 1]. If \code{ci_alpha_i} is \code{NA}, existing alpha values are preserved.}

\item{omit_series}{A vector containing the column names of the climate series NOT to be plotted if \code{series = NULL}. This has been added to leave out the Keeling Curve data, but it could later include other regularly updated time-series data, related to climate, which wouldn't normally be plotted in the same window.}

\item{...}{Passed to \code{\link[zoo]{plot.zoo}} for the models.}
}
\description{
Readily displays annotated climate time series (usually temperature) on a common plot along with CMIP Phase 3 or Phase 5 model projections.
}
\examples{
\dontrun{
inst <- get_climate_data(download=FALSE, baseline=TRUE)
cmip5 <- get_models_data(ensemble="cmip5", baseline=1981:2010, save=FALSE) # Load CMIP5 data from combined R data set.

## Plot all CMIP5 scenario realizations, no instrumental temperature series.
plot_models_and_climate_data(inst, cmip5, series=NA, scenario=NULL, start=1950, end=2100, ma=12, baseline=NULL,
  center_fun="mean", smooth_envelope=TRUE, col_m_mean="red", ylim=c(-1, 5))

## CMIP5 RCP 4.5 scenario realizations compared to the GISTEMP land+SST series, both with 95\% confidence intervals.
series <- c("GISTEMP")
plot_models_and_climate_data(inst, cmip5, series=series, scenario="RCP 4.5", start=1880, end=2020, ma=12, ma_i=12,
  baseline=1951:1980, center_fun="mean", smooth_envelope=TRUE, envelope_type="quantiles", envelope_text="quantiles",
  ylim=c(-1.0, 1.5), conf_int_i=FALSE, col_i_fun="topo.colors", col_i_fun...=NULL)

## Same plot, different baseline:
plot_models_and_climate_data(inst, cmip5, series=series, scenario="RCP 4.5", start=1880, end=2020, ma=12, ma_i=12,
  baseline=1981:2010, center_fun="mean", smooth_envelope=TRUE, envelope_type="quantiles", envelope_text="quantiles",
  ylim=c(-1.5, 1.0), conf_int_i=TRUE, col_i_fun=function(...) "red")
}
}

